backend/pyproject.toml
[project]
name = "backends"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography==46.0.3",
    "fastapi==0.128.0",
    "psycopg2-binary==2.9.11",
    "pydantic[email]==2.12.5",
    "pyjwt[crypto]==2.10.1",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "sqlmodel==0.0.31",
    "uvicorn[standard]==0.40.0",
]
